Update attribution order on internal pages
