+++ app/routes/__docs/$/$.tsx
@@ -92,8 +92,6 @@ export default () => {
         sidebarItems={data.sidebar}
         editPageUrl={page.origin.html_url || undefined}
         toc={page.toc}
-        readTime={page.readTime}
-        attributionData={page.origin.attributionData}
       >
         <MDXContent />
       </InternalPageContent>
